+++ src/main/kotlin/com/faimdata/booker/controllers/ProviderController.kt
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 
 @RestController
-@RequestMapping("/api/providers")
+@RequestMapping("/providers")
 class ProviderController(
         @Autowired private val providerRepo: ProviderRepo
 ) {
@@ -21,16 +21,15 @@ class ProviderController(
         return "Saved provider..."
     }
 
-    @PutMapping("/update/{id}")
-    fun updateProvider(@PathVariable id: String, @RequestBody provider: Provider): String {
-        providerRepo.deleteById(id)
-        providerRepo.save(provider)
-        return "Updated provider..."
-    }
-
     @DeleteMapping("/delete/{id}")
     fun deleteProvider(@PathVariable id: String): String {
+        if (providerRepo.containsAppointments(providerId = id)) {
+            return "Provider cannot be deleted... has upcoming appointments"
+        }
+        // TODO: Should delete attached appointments
         providerRepo.deleteById(id)
         return "Deleted provider..."
     }
+
+    // TODO: if we want to add an update function, we should also update the clinic and appointments attached to current id
 }
